+++ benches/over.rs
@@ -8,11 +8,11 @@ use criterion::Criterion;
 fn gray_over(c: &mut Criterion, sz: u32) {
     let s = format!("gray_over_{}", sz);
     c.bench_function(&s, move |b| {
-        let mut r = Raster::<Gray8>::new(sz, sz);
-        let mut m = Raster::<Alpha8>::new(sz, sz);
-        m.set_pixel(0, 0, Alpha8::new(255));
-        m.set_pixel(sz - 1, sz - 1, Alpha8::new(128));
-        b.iter(|| r.mask_over(&m, 0, 0, Gray8::new(100)))
+        let mut r = Raster::<Gray<Cu8>>::new(sz, sz);
+        let mut m = Raster::<Alpha<Cu8>>::new(sz, sz);
+        m.set_pixel(0, 0, Alpha::<Cu8>::new(255.into()));
+        m.set_pixel(sz - 1, sz - 1, Alpha::<Cu8>::new(128.into()));
+        b.iter(|| r.mask_over(&m, 0, 0, Gray::<Cu8>::new(100.into())))
     });
 }
 
@@ -31,11 +31,12 @@ fn gray_over_512(c: &mut Criterion) {
 fn rgba_over(c: &mut Criterion, sz: u32) {
     let s = format!("rgba_over_{}", sz);
     c.bench_function(&s, move |b| {
-        let mut r = Raster::<Rgba8>::new(sz, sz);
-        let mut m = Raster::<Alpha8>::new(sz, sz);
-        m.set_pixel(0, 0, Alpha8::new(255));
-        m.set_pixel(sz - 1, sz - 1, Alpha8::new(128));
-        b.iter(|| r.mask_over(&m, 0, 0, Rgba8::new(100, 50, 150, 255)))
+        let mut r = Raster::<Rgba<Cu8>>::new(sz, sz);
+        let mut m = Raster::<Alpha<Cu8>>::new(sz, sz);
+        let rgba = Rgba::new(100.into(), 50.into(), 150.into(), 255.into());
+        m.set_pixel(0, 0, Alpha::<Cu8>::new(255.into()));
+        m.set_pixel(sz - 1, sz - 1, Alpha::<Cu8>::new(128.into()));
+        b.iter(|| r.mask_over(&m, 0, 0, rgba))
     });
 }